+++ HISTORY.rst
@@ -4,6 +4,7 @@ Changelog
 0.6.3 (unreleased)
 ------------------
 - Minor word tokenization fix: Words that stem from a contraction will still have an apostrophe, e.g. ``"Let's" => ["Let", "'s"]``. This makes it easier to identify contractions.
+- Fix bug with comparing blobs to strings.
 
 0.6.2 (2013-09-05)
 ------------------